requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
docutils==0.21.2
idna==3.7
pillow==10.3.0
requests==2.32.2
tomli_w==1.0.0
urllib3==2.2.1
